PITERTUL. Ajustes de limpieza y empaquetado
Former-commit-id: 115e0a4eed03ae5051101c5109564ae2990b69bc

+++ sql/tool/run-single-script.sh
@@ -199,7 +199,9 @@ nombreFicheroSinDir=`basename $FICHERO`
 nombreSinDirSinExt=${nombreFicheroSinDir%%.*}
 nombreSetEnv=setEnv_${nombreSinDirSinExt}.sh
 
-rm -f $BASEDIR/tmp/*$nombreSinDirSinExt* 
+if [[ $PACKAGE == 0 ]]; then
+    rm -f $BASEDIR/tmp/*$nombreSinDirSinExt* 
+fi
 
 if [[ ! $nombreFicheroSinDir =~ ^D[MD]L_[0-9]+_[^_]+_[^\.]+\.sql$ ]] ; then
     echo ""
